move open file message listener to viewer pane
